Move pipelines components to lib folder, merge API with the rest, do extra cleanup
