--- api/requirements.txt
@@ -1,6 +1,0 @@
-futures==3.0.5
-Jinja2==2.8
-PyYAML==3.11
-requests==2.9.1
-sh==1.11
-tornado==4.3
