--- api/workspace/sample_pipe2/50ec8b7b-ea4f-4df3-9362-1912b971c406/status.json
@@ -1,3 +1,0 @@
-{
-  "status": "processing"
-}
