+++ lib/pipeline/exceptions.py
