+++ lib/plugin/base_plugin.py
