+++ lib/plugins/file_logger.py
@@ -1,7 +1,7 @@
 import json
 import logging
-from pipelineplugins.stdout_logger import StdoutLogger
-from pluginworm.exceptions import PluginError
+from pipelines.plugins.stdout_logger import StdoutLogger
+from pipelines.plugin.exceptions import PluginError
 
 RETRY_COUNT = 2
