--- tasker/sampleprj/sample_pipe.yaml
@@ -1,17 +1,0 @@
-vars:
-  workspace: '/Users/juha/work/getpipeline/.venv'
-  virtualenv_dir: '/Users/juha/work/getpipeline/.venv'
-
-configuration:
-  webhook_url: 'https://hooks.slack.com/services/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'
-
-tasks:
- - type: bash
-   cmd: "cd {{workspace}} && echo 'Starting' >> logfile.txt"
- - type: python
-   virtualenv: "{{virtualenv_dir}}"
-   workdir: "{{workspace}}"
-   script: test_script.py
-
-plugins:
-  - webhook_logger
