--- tasker/scripts/pline
@@ -1,44 +1,0 @@
-#!/usr/bin/env python
-
-"""
-Usage:
-    pipeline (--help)
-    pipeline <pipeline-yaml> [--verbose]
-
-Help:
-    pipeline                                          Run a pipeline
-"""
-from os import path
-from docopt import docopt
-
-import logging
-
-from pipelineworm.exceptions import PipelineError
-from pipelineworm.pipeline import Pipeline
-from pluginworm.utils import setup_logging
-
-
-def main():
-    args = docopt(__doc__, version="Pipelines 0.1.0")
-
-    configfile = path.abspath(args['<pipeline-yaml>']) if args['<pipeline-yaml>'] else None
-    verbose = int(args['--verbose']) if args['--verbose'] else False
-
-    if verbose:
-        setup_logging(logging.DEBUG)
-    else:
-        setup_logging(logging.INFO)
-
-    log = logging.getLogger()
-
-    try:
-        pipe = Pipeline.from_yaml(configfile)
-        pipe.run()
-    except PipelineError as e:
-        log.error(e.message)
-        exit(1)
-
-if __name__ == '__main__':
-    main()
-
-
