--- tasker/test/test_script.py
@@ -1,3 +1,0 @@
-import os
-import sh
-print 'This is test script'
